trying to debug env vars
